Move InstitutionListFilters to components
